requirements.txt
youtube-dl
spotipy
python-dotenv